🐛 Move position of player when moved to the b2 using elevator
We moved the position because of the previous position was next to the wall. In
mobile, it made character not be able to move.

We tried to move the character slightly upward. However, it triggers the
elevator pannel. So we moved character away from elevator. It looks awkward but
it is the only easy solution.

Fixes CUB-50